return delete index elasticsearch when seed db

+++ database/seeds/ProductTableSeeder.php
@@ -18,6 +18,8 @@ class ProductTableSeeder extends Seeder
      */
     public function run()
     {
+    	Product::truncate();
+
         // delete Elastic Product Index
         $param = [
             'index' => Config::get('constants.elasticsearch.product.index')
@@ -25,9 +27,9 @@ class ProductTableSeeder extends Seeder
         $client = ClientBuilder::create()->build();
 
         // check index exists before delete
-        /*if ($client->indices()->exists($param)) {
+        if ($client->indices()->exists($param)) { dd('a');
             $client->indices()->delete($param);
-        }*/
+        }
 
         // add data to mongo db product table
     	$productRepository = app(ProductRepositoryInterface::class);
